refactor: extract QR code forms and fix mobile responsiveness
## QR Code Generator Refactoring
- Extract 17 form components to components/qr-code/forms/
  - BasicForms.tsx: URL, Text, Email, Phone, SMS
  - WiFiForm.tsx: WiFi network configuration
  - VCardForm.tsx: Contact card with address fields
  - LocationEventForms.tsx: Location and calendar event
  - SocialForms.tsx: Twitter, YouTube, Facebook, App Store
  - CryptoForms.tsx: Bitcoin, Ethereum, Cardano, Solana
- Reduce qr-code-generator.tsx from 3,263 to 2,723 lines (~540 lines)
- Add shared FormProps type in forms/types.ts
- Update barrel exports in components/qr-code/index.ts

## Mobile Responsiveness Fixes
- ElementToolbar: Make color picker responsive on 320px screens
- ReferrerList: Add mobile-friendly truncation (120px -> 200px)
- url-shortener: Smaller button min-width on mobile
- index.tsx: Smaller share dropdown on mobile
