+++ components/qr-code/forms/index.ts
@@ -1,0 +1,34 @@
+/**
+ * QR Code Forms - Barrel Export
+ *
+ * Re-exports all QR code form components for convenient imports
+ */
+
+// Types
+export type { FormProps } from "./types";
+
+// Basic Forms
+export { URLForm, TextForm, EmailForm, PhoneForm, SMSForm } from "./BasicForms";
+
+// Network/Contact Forms
+export { WiFiForm } from "./WiFiForm";
+export { VCardForm } from "./VCardForm";
+
+// Location/Event Forms
+export { LocationForm, EventForm } from "./LocationEventForms";
+
+// Social Media Forms
+export {
+  TwitterForm,
+  YouTubeForm,
+  FacebookForm,
+  AppStoreForm,
+} from "./SocialForms";
+
+// Crypto Payment Forms
+export {
+  BitcoinForm,
+  EthereumForm,
+  CardanoForm,
+  SolanaForm,
+} from "./CryptoForms";
