+++ components/qr-code/index.ts
@@ -6,3 +6,6 @@ export { ColorPickerSection } from "./ColorPickerSection";
 export { BatchStylePanel } from "./BatchStylePanel";
 export { InputField, SelectField, CheckboxField } from "./FormFields";
 export { QR_TYPES, DEFAULT_STYLE, LOCALE_MAP } from "./constants";
+
+// QR Code Form Components
+export * from "./forms";
