+++ pages/tools/qr-code-generator.tsx
@@ -65,6 +65,25 @@ import {
   InputField,
   SelectField,
   CheckboxField,
+  // Form components
+  URLForm,
+  TextForm,
+  EmailForm,
+  PhoneForm,
+  SMSForm,
+  WiFiForm,
+  VCardForm,
+  LocationForm,
+  EventForm,
+  TwitterForm,
+  YouTubeForm,
+  FacebookForm,
+  AppStoreForm,
+  BitcoinForm,
+  EthereumForm,
+  CardanoForm,
+  SolanaForm,
+  type FormProps,
 } from "@/components/qr-code";
 
 // ============================================================================
@@ -192,564 +211,6 @@ interface HistoryItem {
 }
 
 type DownloadFormat = "png" | "svg" | "jpg";
-
-// ============================================================================
-// Input Form Components by Type
-// ============================================================================
-
-interface FormProps {
-  data: Record<string, unknown>;
-  updateField: (field: string, value: unknown) => void;
-  t: (key: string) => string;
-}
-
-function URLForm({ data, updateField, t }: FormProps) {
-  return (
-    <InputField
-      label={t("tools.qrCode.urlLabel")}
-      value={(data.url as string) || ""}
-      onChange={(v: string) => updateField("url", v)}
-      placeholder={t("tools.qrCode.urlPlaceholder")}
-      type="url"
-      required
-    />
-  );
-}
-
-function TextForm({ data, updateField, t }: FormProps) {
-  return (
-    <InputField
-      label={t("tools.qrCode.textLabel")}
-      value={(data.text as string) || ""}
-      onChange={(v: string) => updateField("text", v)}
-      placeholder={t("tools.qrCode.textPlaceholder")}
-      multiline
-      rows={4}
-      required
-    />
-  );
-}
-
-function EmailForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <InputField
-        label={t("tools.qrCode.emailLabel")}
-        value={(data.email as string) || ""}
-        onChange={(v: string) => updateField("email", v)}
-        placeholder={t("tools.qrCode.emailPlaceholder")}
-        type="email"
-        required
-      />
-      <InputField
-        label={t("tools.qrCode.emailSubject")}
-        value={(data.subject as string) || ""}
-        onChange={(v: string) => updateField("subject", v)}
-        placeholder={t("tools.qrCode.emailSubjectPlaceholder")}
-      />
-      <InputField
-        label={t("tools.qrCode.emailBody")}
-        value={(data.body as string) || ""}
-        onChange={(v: string) => updateField("body", v)}
-        placeholder={t("tools.qrCode.emailBodyPlaceholder")}
-        multiline
-        rows={3}
-      />
-    </div>
-  );
-}
-
-function PhoneForm({ data, updateField, t }: FormProps) {
-  return (
-    <InputField
-      label={t("tools.qrCode.phoneLabel")}
-      value={(data.phone as string) || ""}
-      onChange={(v: string) => updateField("phone", v)}
-      placeholder={t("tools.qrCode.phonePlaceholder")}
-      type="tel"
-      required
-    />
-  );
-}
-
-function SMSForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <InputField
-        label={t("tools.qrCode.smsPhone")}
-        value={(data.phone as string) || ""}
-        onChange={(v: string) => updateField("phone", v)}
-        placeholder={t("tools.qrCode.phonePlaceholder")}
-        type="tel"
-        required
-      />
-      <InputField
-        label={t("tools.qrCode.smsMessage")}
-        value={(data.message as string) || ""}
-        onChange={(v: string) => updateField("message", v)}
-        placeholder={t("tools.qrCode.smsMessagePlaceholder")}
-        multiline
-        rows={3}
-      />
-    </div>
-  );
-}
-
-function WiFiForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <InputField
-        label={t("tools.qrCode.wifiSSID")}
-        value={(data.ssid as string) || ""}
-        onChange={(v: string) => updateField("ssid", v)}
-        placeholder={t("tools.qrCode.wifiSSIDPlaceholder")}
-        required
-      />
-      <InputField
-        label={t("tools.qrCode.wifiPassword")}
-        value={(data.password as string) || ""}
-        onChange={(v: string) => updateField("password", v)}
-        placeholder={t("tools.qrCode.wifiPasswordPlaceholder")}
-        type="password"
-      />
-      <SelectField
-        label={t("tools.qrCode.wifiEncryption")}
-        value={(data.encryption as string) || "WPA"}
-        onChange={(v: string) => updateField("encryption", v)}
-        options={[
-          { value: "WPA", label: "WPA/WPA2/WPA3" },
-          { value: "WEP", label: "WEP" },
-          { value: "nopass", label: t("tools.qrCode.wifiNoPassword") },
-        ]}
-      />
-      <CheckboxField
-        label={t("tools.qrCode.wifiHidden")}
-        checked={(data.hidden as boolean) || false}
-        onChange={(v: boolean) => updateField("hidden", v)}
-      />
-    </div>
-  );
-}
-
-function VCardForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-        <InputField
-          label={t("tools.qrCode.vcardFirstName")}
-          value={(data.firstName as string) || ""}
-          onChange={(v: string) => updateField("firstName", v)}
-          placeholder={t("tools.qrCode.vcardFirstNamePlaceholder")}
-          required
-        />
-        <InputField
-          label={t("tools.qrCode.vcardLastName")}
-          value={(data.lastName as string) || ""}
-          onChange={(v: string) => updateField("lastName", v)}
-          placeholder={t("tools.qrCode.vcardLastNamePlaceholder")}
-        />
-      </div>
-      <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-        <InputField
-          label={t("tools.qrCode.vcardOrganization")}
-          value={(data.organization as string) || ""}
-          onChange={(v: string) => updateField("organization", v)}
-          placeholder={t("tools.qrCode.vcardOrganizationPlaceholder")}
-        />
-        <InputField
-          label={t("tools.qrCode.vcardTitle")}
-          value={(data.title as string) || ""}
-          onChange={(v: string) => updateField("title", v)}
-          placeholder={t("tools.qrCode.vcardTitlePlaceholder")}
-        />
-      </div>
-      <InputField
-        label={t("tools.qrCode.vcardEmail")}
-        value={(data.email as string) || ""}
-        onChange={(v: string) => updateField("email", v)}
-        placeholder={t("tools.qrCode.emailPlaceholder")}
-        type="email"
-      />
-      <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-        <InputField
-          label={t("tools.qrCode.vcardPhone")}
-          value={(data.phone as string) || ""}
-          onChange={(v: string) => updateField("phone", v)}
-          placeholder={t("tools.qrCode.phonePlaceholder")}
-          type="tel"
-        />
-        <InputField
-          label={t("tools.qrCode.vcardMobile")}
-          value={(data.mobile as string) || ""}
-          onChange={(v: string) => updateField("mobile", v)}
-          placeholder={t("tools.qrCode.vcardMobilePlaceholder")}
-          type="tel"
-        />
-      </div>
-      <InputField
-        label={t("tools.qrCode.vcardWebsite")}
-        value={(data.website as string) || ""}
-        onChange={(v: string) => updateField("website", v)}
-        placeholder={t("tools.qrCode.urlPlaceholder")}
-        type="url"
-      />
-      <div className="border-t border-zinc-200 dark:border-dark-border pt-4 mt-4">
-        <p className="text-xs text-zinc-500 dark:text-zinc-400 mb-3">
-          {t("tools.qrCode.vcardAddressOptional")}
-        </p>
-        <div className="space-y-3">
-          <InputField
-            label={t("tools.qrCode.vcardStreet")}
-            value={(data.street as string) || ""}
-            onChange={(v: string) => updateField("street", v)}
-            placeholder={t("tools.qrCode.vcardStreetPlaceholder")}
-          />
-          <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-            <InputField
-              label={t("tools.qrCode.vcardCity")}
-              value={(data.city as string) || ""}
-              onChange={(v: string) => updateField("city", v)}
-              placeholder={t("tools.qrCode.vcardCityPlaceholder")}
-            />
-            <InputField
-              label={t("tools.qrCode.vcardState")}
-              value={(data.state as string) || ""}
-              onChange={(v: string) => updateField("state", v)}
-              placeholder={t("tools.qrCode.vcardStatePlaceholder")}
-            />
-          </div>
-          <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-            <InputField
-              label={t("tools.qrCode.vcardZip")}
-              value={(data.zip as string) || ""}
-              onChange={(v: string) => updateField("zip", v)}
-              placeholder={t("tools.qrCode.vcardZipPlaceholder")}
-            />
-            <InputField
-              label={t("tools.qrCode.vcardCountry")}
-              value={(data.country as string) || ""}
-              onChange={(v: string) => updateField("country", v)}
-              placeholder={t("tools.qrCode.vcardCountryPlaceholder")}
-            />
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function LocationForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <p className="text-xs text-zinc-500 dark:text-zinc-400">
-        {t("tools.qrCode.locationHint")}
-      </p>
-      <InputField
-        label={t("tools.qrCode.locationQuery")}
-        value={(data.query as string) || ""}
-        onChange={(v: string) => updateField("query", v)}
-        placeholder={t("tools.qrCode.locationQueryPlaceholder")}
-      />
-      <div className="flex items-center gap-4 my-3">
-        <div className="flex-1 h-px bg-zinc-200 dark:bg-zinc-700" />
-        <span className="text-xs text-zinc-400">{t("tools.qrCode.or")}</span>
-        <div className="flex-1 h-px bg-zinc-200 dark:bg-zinc-700" />
-      </div>
-      <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-        <InputField
-          label={t("tools.qrCode.locationLatitude")}
-          value={(data.latitude as string) || ""}
-          onChange={(v: string) => updateField("latitude", v)}
-          placeholder="e.g., 40.7128"
-        />
-        <InputField
-          label={t("tools.qrCode.locationLongitude")}
-          value={(data.longitude as string) || ""}
-          onChange={(v: string) => updateField("longitude", v)}
-          placeholder="e.g., -74.0060"
-        />
-      </div>
-    </div>
-  );
-}
-
-function EventForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <InputField
-        label={t("tools.qrCode.eventTitle")}
-        value={(data.title as string) || ""}
-        onChange={(v: string) => updateField("title", v)}
-        placeholder={t("tools.qrCode.eventTitlePlaceholder")}
-        required
-      />
-      <InputField
-        label={t("tools.qrCode.eventLocation")}
-        value={(data.location as string) || ""}
-        onChange={(v: string) => updateField("location", v)}
-        placeholder={t("tools.qrCode.eventLocationPlaceholder")}
-      />
-      <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-        <InputField
-          label={t("tools.qrCode.eventStartDate")}
-          value={(data.startDate as string) || ""}
-          onChange={(v: string) => updateField("startDate", v)}
-          type="date"
-          required
-        />
-        <InputField
-          label={t("tools.qrCode.eventStartTime")}
-          value={(data.startTime as string) || ""}
-          onChange={(v: string) => updateField("startTime", v)}
-          type="time"
-          required
-        />
-      </div>
-      <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-        <InputField
-          label={t("tools.qrCode.eventEndDate")}
-          value={(data.endDate as string) || ""}
-          onChange={(v: string) => updateField("endDate", v)}
-          type="date"
-        />
-        <InputField
-          label={t("tools.qrCode.eventEndTime")}
-          value={(data.endTime as string) || ""}
-          onChange={(v: string) => updateField("endTime", v)}
-          type="time"
-        />
-      </div>
-      <InputField
-        label={t("tools.qrCode.eventDescription")}
-        value={(data.description as string) || ""}
-        onChange={(v: string) => updateField("description", v)}
-        placeholder={t("tools.qrCode.eventDescriptionPlaceholder")}
-        multiline
-        rows={3}
-      />
-    </div>
-  );
-}
-
-function TwitterForm({ data, updateField, t }: FormProps) {
-  return (
-    <InputField
-      label={t("tools.qrCode.twitterUsername")}
-      value={(data.username as string) || ""}
-      onChange={(v: string) => updateField("username", v)}
-      placeholder={t("tools.qrCode.twitterPlaceholder")}
-      required
-    />
-  );
-}
-
-function YouTubeForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <InputField
-        label={t("tools.qrCode.youtubeVideo")}
-        value={(data.videoId as string) || ""}
-        onChange={(v: string) => updateField("videoId", v)}
-        placeholder={t("tools.qrCode.youtubeVideoPlaceholder")}
-      />
-      <div className="flex items-center gap-4 my-3">
-        <div className="flex-1 h-px bg-zinc-200 dark:bg-zinc-700" />
-        <span className="text-xs text-zinc-400">{t("tools.qrCode.or")}</span>
-        <div className="flex-1 h-px bg-zinc-200 dark:bg-zinc-700" />
-      </div>
-      <InputField
-        label={t("tools.qrCode.youtubeChannel")}
-        value={(data.channelId as string) || ""}
-        onChange={(v: string) => updateField("channelId", v)}
-        placeholder={t("tools.qrCode.youtubeChannelPlaceholder")}
-      />
-    </div>
-  );
-}
-
-function FacebookForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <InputField
-        label={t("tools.qrCode.facebookUsername")}
-        value={(data.username as string) || ""}
-        onChange={(v: string) => updateField("username", v)}
-        placeholder={t("tools.qrCode.facebookPlaceholder")}
-      />
-      <p className="text-xs text-zinc-500 dark:text-zinc-400">
-        {t("tools.qrCode.facebookHint")}
-      </p>
-    </div>
-  );
-}
-
-function BitcoinForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <InputField
-        label={t("tools.qrCode.bitcoinAddress")}
-        value={(data.address as string) || ""}
-        onChange={(v: string) => updateField("address", v)}
-        placeholder={t("tools.qrCode.bitcoinAddressPlaceholder")}
-        required
-      />
-      <InputField
-        label={t("tools.qrCode.bitcoinAmount")}
-        value={(data.amount as string) || ""}
-        onChange={(v: string) => updateField("amount", v)}
-        placeholder={t("tools.qrCode.bitcoinAmountPlaceholder")}
-      />
-      <InputField
-        label={t("tools.qrCode.bitcoinLabel")}
-        value={(data.label as string) || ""}
-        onChange={(v: string) => updateField("label", v)}
-        placeholder={t("tools.qrCode.bitcoinLabelPlaceholder")}
-      />
-      <InputField
-        label={t("tools.qrCode.bitcoinMessage")}
-        value={(data.message as string) || ""}
-        onChange={(v: string) => updateField("message", v)}
-        placeholder={t("tools.qrCode.bitcoinMessagePlaceholder")}
-        multiline
-        rows={2}
-      />
-    </div>
-  );
-}
-
-function EthereumForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <InputField
-        label={t("tools.qrCode.ethereumAddress")}
-        value={(data.address as string) || ""}
-        onChange={(v: string) => updateField("address", v)}
-        placeholder={t("tools.qrCode.ethereumAddressPlaceholder")}
-        required
-      />
-      <InputField
-        label={t("tools.qrCode.ethereumAmount")}
-        value={(data.amount as string) || ""}
-        onChange={(v: string) => updateField("amount", v)}
-        placeholder={t("tools.qrCode.ethereumAmountPlaceholder")}
-      />
-      <InputField
-        label={t("tools.qrCode.ethereumToken")}
-        value={(data.tokenAddress as string) || ""}
-        onChange={(v: string) => updateField("tokenAddress", v)}
-        placeholder={t("tools.qrCode.ethereumTokenPlaceholder")}
-      />
-      <SelectField
-        label={t("tools.qrCode.ethereumNetwork")}
-        value={(data.chainId as string) || ""}
-        onChange={(v: string) => updateField("chainId", v)}
-        options={[
-          { value: "", label: "Mainnet (default)" },
-          { value: "1", label: "Ethereum Mainnet" },
-          { value: "137", label: "Polygon" },
-          { value: "56", label: "BNB Smart Chain" },
-          { value: "42161", label: "Arbitrum One" },
-          { value: "10", label: "Optimism" },
-        ]}
-      />
-      <p className="text-xs text-zinc-500 dark:text-zinc-400">
-        {t("tools.qrCode.ethereumHint")}
-      </p>
-    </div>
-  );
-}
-
-function CardanoForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <InputField
-        label={t("tools.qrCode.cardanoAddress")}
-        value={(data.address as string) || ""}
-        onChange={(v: string) => updateField("address", v)}
-        placeholder={t("tools.qrCode.cardanoAddressPlaceholder")}
-        required
-      />
-      <InputField
-        label={t("tools.qrCode.cardanoAmount")}
-        value={(data.amount as string) || ""}
-        onChange={(v: string) => updateField("amount", v)}
-        placeholder={t("tools.qrCode.cardanoAmountPlaceholder")}
-      />
-      <p className="text-xs text-zinc-500 dark:text-zinc-400">
-        {t("tools.qrCode.cardanoHint")}
-      </p>
-    </div>
-  );
-}
-
-function SolanaForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <InputField
-        label={t("tools.qrCode.solanaAddress")}
-        value={(data.address as string) || ""}
-        onChange={(v: string) => updateField("address", v)}
-        placeholder={t("tools.qrCode.solanaAddressPlaceholder")}
-        required
-      />
-      <InputField
-        label={t("tools.qrCode.solanaAmount")}
-        value={(data.amount as string) || ""}
-        onChange={(v: string) => updateField("amount", v)}
-        placeholder={t("tools.qrCode.solanaAmountPlaceholder")}
-      />
-      <InputField
-        label={t("tools.qrCode.solanaLabel")}
-        value={(data.label as string) || ""}
-        onChange={(v: string) => updateField("label", v)}
-        placeholder={t("tools.qrCode.solanaLabelPlaceholder")}
-      />
-      <InputField
-        label={t("tools.qrCode.solanaMessage")}
-        value={(data.message as string) || ""}
-        onChange={(v: string) => updateField("message", v)}
-        placeholder={t("tools.qrCode.solanaMessagePlaceholder")}
-        multiline
-        rows={2}
-      />
-      <p className="text-xs text-zinc-500 dark:text-zinc-400">
-        {t("tools.qrCode.solanaHint")}
-      </p>
-    </div>
-  );
-}
-
-function AppStoreForm({ data, updateField, t }: FormProps) {
-  return (
-    <div className="space-y-4">
-      <SelectField
-        label={t("tools.qrCode.appstorePlatform")}
-        value={(data.platform as string) || "ios"}
-        onChange={(v: string) => updateField("platform", v)}
-        options={[
-          { value: "ios", label: "iOS (App Store)" },
-          { value: "android", label: "Android (Google Play)" },
-        ]}
-      />
-      <InputField
-        label={t("tools.qrCode.appstoreId")}
-        value={(data.appId as string) || ""}
-        onChange={(v: string) => updateField("appId", v)}
-        placeholder={
-          (data.platform as string) === "android"
-            ? t("tools.qrCode.appstoreAndroidPlaceholder")
-            : t("tools.qrCode.appstoreIosPlaceholder")
-        }
-        required
-      />
-      <p className="text-xs text-zinc-500 dark:text-zinc-400">
-        {(data.platform as string) === "android"
-          ? t("tools.qrCode.appstoreAndroidHint")
-          : t("tools.qrCode.appstoreIosHint")}
-      </p>
-    </div>
-  );
-}
 
 // ============================================================================
 // Main Component
